# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vehicles": {
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create vehicle for a driver",
        "tags": [
          "Vehicles"
        ]
      },
      "get": {
        "operationId": "VehiclesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Vehicles fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all vehicles (admin)",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/driver/{driverId}": {
      "get": {
        "operationId": "VehiclesController_findByDriverId",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "description": "Driver UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver vehicles fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List vehicles for a driver",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get vehicle by ID",
        "tags": [
          "Vehicles"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update vehicle details",
        "tags": [
          "Vehicles"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vehicle deleted successfully."
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Delete a vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{id}/deactivate": {
      "patch": {
        "operationId": "VehiclesController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle deactivated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{id}/activate": {
      "patch": {
        "operationId": "VehiclesController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle activated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Activate a vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ark Rides API",
    "description": "Ride-hailing backend for Ark Rides: rider and driver identity (Privy + email), ride lifecycle, fare ledger, driver wallet, emergency SOS and off-app booking channels.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "driverId": {
            "type": "string",
            "example": "123e4567-e89b-42d3-a456-426614174000",
            "description": "Admin only. Drivers must omit this — the owner is taken from the access token."
          },
          "type": {
            "type": "string",
            "enum": [
              "keke",
              "bike",
              "car",
              "courier"
            ],
            "example": "keke"
          },
          "plateNumber": {
            "type": "string",
            "example": "ABC-123XY"
          },
          "color": {
            "type": "string",
            "example": "Yellow"
          },
          "model": {
            "type": "string",
            "example": "Bajaj RE"
          },
          "year": {
            "type": "number",
            "example": 2024,
            "minimum": 1990
          }
        },
        "required": [
          "type",
          "plateNumber",
          "color",
          "model",
          "year"
        ]
      },
...
      "ApiSuccessEnvelope": {
        "type": "object",
        "description": "Returned for every 2xx that carries a body.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Request successful"
          },
          "data": {
            "description": "The payload. Its shape is per-endpoint.",
            "nullable": true
          },
          "meta": {
            "type": "object",
            "description": "Present on list endpoints.",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 20
              },
              "total": {
                "type": "integer",
                "example": 41
              },
              "totalPages": {
                "type": "integer",
                "example": 3
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data",
          "timestamp"
        ]
      }
    }
  }
}